+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.6.47] - 2021-7-13
+
+###Added
+- update media transcode api
+- add media transcode demo
+
 ## [5.6.46] - 2021-7-8
 
 ###Added